requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
backcall==0.2.0
billiard==4.1.0
boto3==1.28.32
botocore==1.31.32
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
dnspython==2.4.2
drf-yasg==1.21.7
executing==1.2.0
flower==2.0.1
humanize==4.8.0
hvac==1.1.1
idna==3.4
inflection==0.5.1
install==1.3.5
ipython==8.12.2
jedi==0.19.0
jmespath==1.0.1
kombu==5.3.1
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyhcl==0.4.4
PyJWT==2.7.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
s3transfer==0.6.2
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
starkbank-ecdsa==2.2.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
